+++ changelog.md
@@ -20,6 +20,7 @@
 * [3792](https://github.com/zeta-chain/node/pull/3792) - add compliance check for Sui inbound and outbound
 * [3790](https://github.com/zeta-chain/node/pull/3790) - integrate execute revert
 * [3797](https://github.com/zeta-chain/node/pull/3797) - integrate execute SPL revert
+* [3807](https://github.com/zeta-chain/node/pull/3807) - integrate ZEVM to Solana call
 
 ### Refactor